[Gold IV] Title: 인문예술탐사주간, Time: 52 ms, Memory: 3984 KB -BaekjoonHub

+++ 백준/Gold/23829. 인문예술탐사주간/인문예술탐사주간.cc
@@ -9,7 +9,7 @@ int main() {
 
     cin >> N >> Q;
     
-    vector<long long> tree(N);
+    vector<int> tree(N);
     vector<long long> sumTree(N);
     vector<long long> picture(Q);
